one route connected to postgres 🙂

+++ src/types/food.rs
@@ -1,10 +1,11 @@
 use serde::{Deserialize, Serialize};
+use sqlx::Type;
 
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug, Type)]
+#[sqlx(type_name = "food")] 
 pub struct Food {
     pub name: String,
-    pub price: usize,
+    pub price: i32,
     pub ingredient: Option<Vec<String>>,
     pub available: bool,
 }
